+++ engine/client/cl_console.cpp
@@ -254,7 +254,7 @@ static void Con_Dump_f( void )
 				break;
 		}
 		Q_strcat( buffer, bufferlen, Q_NEWLINE );
-		FS_Write( buffer, strlen( buffer ), f );
+		FS_Write( buffer, (int)( strlen( buffer ) ), f );
 	}
 
 	Hunk_FreeTempMemory( buffer );
@@ -837,7 +837,7 @@ static void Con_DrawSolidConsole( float frac ) {
 		currentColorIndex = ColorIndex( COLOR_CYAN );
 		re.SetColor( g_color_table[ currentColorIndex ] );
 
-		i = strlen( download.progress );
+		i = (int)( strlen( download.progress ) );
 		for ( x = 0 ; x < i ; x++ ) 
 		{
 			SCR_DrawSmallChar( ( x + 1 ) * smallchar_width,
