+++ engine/renderer/tr_bsp.cpp
@@ -2245,7 +2245,7 @@ static void R_LoadEntities( const lump_t *l ) {
 
 		// check for remapping of shaders for vertex lighting
 		s = "vertexremapshader";
-		if (!Q_strncmp(keyname, s, strlen(s)) ) {
+		if (!Q_strncmp(keyname, s, (int)( strlen(s) )) ) {
 			char *vs = strchr(value, ';');
 			if (!vs) {
 				ri.Printf( PRINT_WARNING, "WARNING: no semi colon in vertexshaderremap '%s'\n", value );
